requirements.txt
# Core Framework Dependencies
python-dotenv>=1.0.0  # For loading environment variables from .env file
PyYAML>=6.0         # For robustly parsing the config/checklist.yml file

# LLM & AI Dependencies
litellm>=1.30.0     # Excellent library for calling 100+ LLM providers using a unified interface. Implements our factory function automatically.
openai              # Required by LiteLLM for OpenAI/Azure providers.

# Agent-Specific Tooling
# Add libraries here as needed by specific agents.
python-docx         # For Agent 1.100 to create .docx files.
requests            # General purpose HTTP requests, useful for API-based agents.
pillow              # For image generation in Agent 2.200

# Testing Dependencies
pytest>=7.0.0       # The testing framework for running tests.
pytest-mock>=3.0.0  # Fixture for easy mocking of functions and objects.

# Optional Web Framework (if you build a UI)
# fastapi
# uvicorn[standard]
# streamlit
# python-multipart
